requirements.txt
langchain==0.1.17
langchain_anthropic==0.1.11
langchain_core==0.1.51
langchain_google_genai==1.0.3
langchain_groq==0.1.3
python-dotenv==1.0.1

numpy==1.26.4
keras==3.3.3
openai==1.28.1
tensorflow==2.16.1
bentoml==1.2.15
python-dotenv==1.0.1
